1-source-material/reuven-cohen-projects/FACT/examples/arcade-dev/requirements.txt
# Arcade.dev Integration Examples - Python Dependencies

# Core HTTP and async libraries
aiohttp>=3.8.0
asyncio-throttle>=1.0.2
httpx>=0.24.0

# Data handling
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Caching and storage
redis>=4.5.0
diskcache>=5.6.0

# Logging and monitoring
structlog>=23.0.0
prometheus-client>=0.16.0

# Testing utilities
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
respx>=0.20.0

# Code analysis and formatting
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0

# Development utilities
ipython>=8.0.0
rich>=13.0.0
typer>=0.9.0